feat(content-sharing): Launch ContentSharing from a custom button
