+++ src/elements/content-sharing/ContentSharing.js
@@ -1,219 +1,106 @@
 /**
  * @flow
  * @file ContentSharing Element
+ * @description This is the top-level component for ContentSharing. It instantiates the API, which it then
+ * passes to the SharingModal component either immediately (when no custom button is provided) or on
+ * button click (when a custom button is provided).
  * @author Box
  */
 import * as React from 'react';
-import { FormattedMessage } from 'react-intl';
-import type { $AxiosError } from 'axios';
 import API from '../../api';
-import Internationalize from '../common/Internationalize';
-import ErrorMask from '../../components/error-mask/ErrorMask';
-import UnifiedShareModal from '../../features/unified-share-modal';
-import SharedLinkSettingsModal from '../../features/shared-link-settings-modal';
-import SharingNotification from './SharingNotification';
-import usmMessages from '../../features/unified-share-modal/messages';
-import { convertItemResponse, convertUserResponse } from '../../features/unified-share-modal/utils/convertData';
-import { FIELD_ENTERPRISE, FIELD_HOSTNAME, TYPE_FILE, TYPE_FOLDER } from '../../constants';
-import { CONTENT_SHARING_ERRORS, CONTENT_SHARING_ITEM_FIELDS, CONTENT_SHARING_VIEWS } from './constants';
-import { INVITEE_PERMISSIONS } from '../../features/unified-share-modal/constants';
-import contentSharingMessages from './messages';
-import type { ErrorResponseData } from '../../common/types/api';
+import SharingModal from './SharingModal';
+import { CLIENT_NAME_CONTENT_SHARING } from '../../constants';
 import type { ItemType } from '../../common/types/core';
-import type { collaboratorsListType, item as itemFlowType } from '../../features/unified-share-modal/flowTypes';
-import type {
-    ContentSharingItemAPIResponse,
-    ContentSharingSharedLinkType,
-    GetContactsFnType,
-    SendInvitesFnType,
-    SharedLinkUpdateLevelFnType,
-    SharedLinkUpdateSettingsFnType,
-} from './types';
 
 type ContentSharingProps = {
-    api: API,
+    apiHost: string,
+    customButton?: React.Element<any>,
     displayInModal: boolean,
     itemID: string,
     itemType: ItemType,
     language: string,
-    onRequestClose?: () => void,
+    token: string,
 };
 
-function ContentSharing({ api, displayInModal, itemID, itemType, language, onRequestClose }: ContentSharingProps) {
-    const [item, setItem] = React.useState<itemFlowType | null>(null);
-    const [sharedLink, setSharedLink] = React.useState<ContentSharingSharedLinkType | null>(null);
-    const [currentUserID, setCurrentUserID] = React.useState<string | null>(null);
-    const [componentErrorMessage, setComponentErrorMessage] = React.useState<Object | null>(null);
-    const [collaboratorsList, setCollaboratorsList] = React.useState<collaboratorsListType | null>(null);
-    const [onAddLink, setOnAddLink] = React.useState<null | SharedLinkUpdateLevelFnType>(null);
-    const [onRemoveLink, setOnRemoveLink] = React.useState<null | SharedLinkUpdateLevelFnType>(null);
-    const [
-        changeSharedLinkAccessLevel,
-        setChangeSharedLinkAccessLevel,
-    ] = React.useState<null | SharedLinkUpdateLevelFnType>(null);
-    const [
-        changeSharedLinkPermissionLevel,
-        setChangeSharedLinkPermissionLevel,
-    ] = React.useState<null | SharedLinkUpdateLevelFnType>(null);
-    const [onSubmitSettings, setOnSubmitSettings] = React.useState<null | SharedLinkUpdateSettingsFnType>(null);
-    const [currentView, setCurrentView] = React.useState<string>(CONTENT_SHARING_VIEWS.UNIFIED_SHARE_MODAL);
-    const [getContacts, setGetContacts] = React.useState<null | GetContactsFnType>(null);
-    const [sendInvites, setSendInvites] = React.useState<null | SendInvitesFnType>(null);
+const createAPI = (apiHost, itemID, itemType, token) =>
+    new API({
+        apiHost,
+        clientName: CLIENT_NAME_CONTENT_SHARING,
+        id: `${itemType}_${itemID}`,
+        token,
+    });
 
-    // Handle successful GET requests to /files or /folders
-    const handleGetItemSuccess = React.useCallback((itemData: ContentSharingItemAPIResponse) => {
-        const { item: itemFromAPI, sharedLink: sharedLinkFromAPI } = convertItemResponse(itemData);
-        setComponentErrorMessage(null);
-        setItem(itemFromAPI);
-        setSharedLink(sharedLinkFromAPI);
-    }, []);
-
-    // Handle component-level errors
-    const getError = React.useCallback(
-        (error: $AxiosError<Object> | ErrorResponseData) => {
-            let errorObject;
-            if (error.status) {
-                errorObject = contentSharingMessages[CONTENT_SHARING_ERRORS[error.status]];
-            } else if (error.response && error.response.status) {
-                errorObject = contentSharingMessages[CONTENT_SHARING_ERRORS[error.response.status]];
-            } else {
-                errorObject = contentSharingMessages.loadingError;
-            }
-
-            setComponentErrorMessage(errorObject);
-        },
-        [setComponentErrorMessage],
+function ContentSharing({
+    apiHost,
+    customButton,
+    displayInModal,
+    itemID,
+    itemType,
+    language,
+    token,
+}: ContentSharingProps) {
+    const [isOpen, setIsOpen] = React.useState<boolean>(true);
+    const [api, setAPI] = React.useState<API>(createAPI(apiHost, itemID, itemType, token));
+    const [launchButton, setLaunchButton] = React.useState<React.Element<any> | null>(null);
+    const [sharingModalInstance, setSharingModalInstance] = React.useState<React.Element<typeof SharingModal> | null>(
+        customButton ? null : (
+            <SharingModal api={api} displayInModal={false} itemID={itemID} itemType={itemType} language={language} />
+        ),
     );
+
+    // Reset the API if necessary
+    React.useEffect(() => {
+        setAPI(createAPI(apiHost, itemID, itemType, token));
+    }, [apiHost, itemID, itemType, token]);
 
     // Reset state if the API has changed
     React.useEffect(() => {
-        setChangeSharedLinkAccessLevel(null);
-        setChangeSharedLinkPermissionLevel(null);
-        setCollaboratorsList(null);
-        setCurrentUserID(null);
-        setItem(null);
-        setOnAddLink(null);
-        setOnRemoveLink(null);
-        setSharedLink(null);
-    }, []);
+        setSharingModalInstance(null);
+        setLaunchButton(null);
+    }, [api]);
 
-    // Get initial data for the item
     React.useEffect(() => {
-        const getItem = () => {
-            if (itemType === TYPE_FILE) {
-                api.getFileAPI().getFile(itemID, handleGetItemSuccess, getError, {
-                    fields: CONTENT_SHARING_ITEM_FIELDS,
-                });
-            } else if (itemType === TYPE_FOLDER) {
-                api.getFolderAPI().getFolderFields(itemID, handleGetItemSuccess, getError, {
-                    fields: CONTENT_SHARING_ITEM_FIELDS,
-                });
-            }
-        };
-
-        if (!item && !sharedLink) {
-            getItem();
-        }
-    }, [api, getError, handleGetItemSuccess, item, itemID, itemType, sharedLink]);
-
-    // Get initial data for the user
-    React.useEffect(() => {
-        const getUserSuccess = userData => {
-            const { id, userEnterpriseData } = convertUserResponse(userData);
-            setCurrentUserID(id);
-            setSharedLink(prevSharedLink => ({ ...prevSharedLink, ...userEnterpriseData }));
-            setComponentErrorMessage(null);
-        };
-
-        const getUserData = () => {
-            api.getUsersAPI(false).getUser(itemID, getUserSuccess, getError, {
-                params: {
-                    fields: [FIELD_ENTERPRISE, FIELD_HOSTNAME].toString(),
-                },
-            });
-        };
-
-        if (item && sharedLink && !currentUserID) {
-            getUserData();
-        }
-    }, [getError, item, itemID, itemType, sharedLink, currentUserID, api]);
-
-    if (componentErrorMessage) {
-        return <ErrorMask errorHeader={<FormattedMessage {...componentErrorMessage} />} />;
-    }
-
-    // Ensure that all necessary data has been received before rendering child components
-    // "serverURL" is added to sharedLink after the call to the Users API
-    if (!item || !sharedLink || !currentUserID || !sharedLink.serverURL) {
-        return null;
-    }
-
-    const { ownerEmail, ownerID, permissions } = item;
-    const { accessLevel = '', serverURL } = sharedLink;
-    return (
-        <Internationalize language={language} messages={usmMessages}>
-            <>
-                <SharingNotification
-                    accessLevel={accessLevel}
+        const createSharingModalInstance = () => {
+            setIsOpen(true);
+            return (
+                <SharingModal
                     api={api}
-                    collaboratorsList={collaboratorsList}
-                    currentUserID={currentUserID}
-                    getContacts={getContacts}
+                    displayInModal={displayInModal}
                     itemID={itemID}
                     itemType={itemType}
-                    onRequestClose={() => setCurrentView(CONTENT_SHARING_VIEWS.UNIFIED_SHARE_MODAL)}
-                    onSubmitSettings={onSubmitSettings}
-                    ownerEmail={ownerEmail}
-                    ownerID={ownerID}
-                    permissions={permissions}
-                    sendInvites={sendInvites}
-                    serverURL={serverURL}
-                    setChangeSharedLinkAccessLevel={setChangeSharedLinkAccessLevel}
-                    setChangeSharedLinkPermissionLevel={setChangeSharedLinkPermissionLevel}
-                    setGetContacts={setGetContacts}
-                    setCollaboratorsList={setCollaboratorsList}
-                    setItem={setItem}
-                    setOnAddLink={setOnAddLink}
-                    setOnRemoveLink={setOnRemoveLink}
-                    setOnSubmitSettings={setOnSubmitSettings}
-                    setSendInvites={setSendInvites}
-                    setSharedLink={setSharedLink}
+                    language={language}
+                    onRequestClose={() => setIsOpen(false)}
                 />
-                {currentView === CONTENT_SHARING_VIEWS.SHARED_LINK_SETTINGS && (
-                    <SharedLinkSettingsModal
-                        isDirectLinkUnavailableDueToDownloadSettings={false}
-                        isDirectLinkUnavailableDueToAccessPolicy={false}
-                        isDirectLinkUnavailableDueToMaliciousContent={false}
-                        isOpen
-                        item={item}
-                        onRequestClose={() => setCurrentView(CONTENT_SHARING_VIEWS.UNIFIED_SHARE_MODAL)}
-                        onSubmit={onSubmitSettings}
-                        {...sharedLink}
-                    />
-                )}
-                {currentView === CONTENT_SHARING_VIEWS.UNIFIED_SHARE_MODAL && (
-                    <UnifiedShareModal
-                        canInvite={sharedLink.canInvite}
-                        changeSharedLinkAccessLevel={changeSharedLinkAccessLevel}
-                        changeSharedLinkPermissionLevel={changeSharedLinkPermissionLevel}
-                        collaboratorsList={collaboratorsList}
-                        currentUserID={currentUserID}
-                        displayInModal={displayInModal}
-                        getCollaboratorContacts={getContacts}
-                        initialDataReceived
-                        inviteePermissions={INVITEE_PERMISSIONS}
-                        isOpen
-                        item={item}
-                        onAddLink={onAddLink}
-                        onRequestClose={onRequestClose}
-                        onRemoveLink={onRemoveLink}
-                        onSettingsClick={() => setCurrentView(CONTENT_SHARING_VIEWS.SHARED_LINK_SETTINGS)}
-                        sendInvites={sendInvites}
-                        sharedLink={sharedLink}
-                    />
-                )}
-            </>
-        </Internationalize>
+            );
+        };
+
+        // Add an onClick function that instantiates SharingModal to the custom button
+        if (customButton && !launchButton) {
+            setLaunchButton(
+                React.cloneElement(customButton, {
+                    onClick: () => {
+                        return setSharingModalInstance(createSharingModalInstance());
+                    },
+                }),
+            );
+        }
+
+        // If there is no custom button, instantiate SharingModal
+        if (!customButton && !sharingModalInstance) {
+            setSharingModalInstance(createSharingModalInstance());
+        }
+
+        // If the modal is closed, remove the SharingModal instance
+        if (!isOpen) {
+            setSharingModalInstance(null);
+        }
+    }, [api, sharingModalInstance, customButton, displayInModal, isOpen, itemID, itemType, language, launchButton]);
+
+    return (
+        <>
+            {sharingModalInstance}
+            {launchButton}
+        </>
     );
 }
