+++ src/elements/content-sharing/stories/ContentSharing.stories.md
@@ -1,3 +1,3 @@
-`import ContentSharing from 'box-ui-elements/es/elements/unified-share-modal-element';`
+`import ContentSharing from 'box-ui-elements/es/elements/content-sharing';`
 
 The ContentSharing Element extends the <a href="/docs/features-unifiedsharemodal--basic">UnifiedShareModal</a> with Box APIs. The `ErrorMask` appears by default if the required props (token, item ID, and item type) are invalid or missing. To test it with your files or folders, update the values in the <a href="/story/elements-contentsharing--basic">Knobs</a> tab of this story.
